+++ sdk/metric/manual_reader.go
@@ -37,6 +37,8 @@ type ManualReader struct {
 	cardinalityLimitSelector CardinalityLimitSelector
 
 	inst *observ.Instrumentation
+
+	metricFilter metricFilter
 }
 
 // Compile time check the manualReader implements Reader and is comparable.
@@ -49,6 +51,7 @@ func NewManualReader(opts ...ManualReaderOption) *ManualReader {
 		temporalitySelector:      cfg.temporalitySelector,
 		aggregationSelector:      cfg.aggregationSelector,
 		cardinalityLimitSelector: cfg.cardinalityLimitSelector,
+		metricFilter:             cfg.metricFilter,
 	}
 	r.externalProducers.Store(cfg.producers)
 
@@ -94,6 +97,11 @@ func (mr *ManualReader) aggregation(
 // cardinalityLimit returns the cardinality limit for kind.
 func (mr *ManualReader) cardinalityLimit(kind InstrumentKind) (int, bool) {
 	return mr.cardinalityLimitSelector(kind)
+}
+
+// metricFilter returns the MetricFilter configured for this reader.
+func (mr *ManualReader) getMetricFilter() metricFilter {
+	return mr.metricFilter
 }
 
 // Shutdown closes any connections and frees any resources used by the reader.
@@ -190,6 +198,7 @@ type manualReaderConfig struct {
 	aggregationSelector      AggregationSelector
 	cardinalityLimitSelector CardinalityLimitSelector
 	producers                []Producer
+	metricFilter             metricFilter
 }
 
 // newManualReaderConfig returns a manualReaderConfig configured with options.
@@ -200,6 +209,13 @@ func newManualReaderConfig(opts []ManualReaderOption) manualReaderConfig {
 		cardinalityLimitSelector: defaultCardinalityLimitSelector,
 	}
 	for _, opt := range opts {
+		if o, ok := opt.(metricFilter); ok {
+			cfg.metricFilter = o
+			continue
+		}
+		if _, ok := opt.(experimentalOption); ok {
+			continue
+		}
 		cfg = opt.applyManual(cfg)
 	}
 	return cfg